get latest updates into dev
* backend Done

* refactor from oop to functional for react to work

* merchant selection working

* create nextDay button

* item shows in focusWindow when clicked

* correctly show information of focusItem

* convert backend currency from double to int

+++ src/main/kotlin/com/projects/inGameMarketplace/gameLogicService/GameDataDTO.kt
@@ -2,4 +2,5 @@ package com.projects.inGameMarketplace.gameLogicService
 
 import com.projects.inGameMarketplace.itemService.Item
 
-data class GameDataDTO(val inventorySpace: Int, val day: Int, val money: Double, val inventoryItems: MutableList<Pair<Item, Int>>)
+data class GameDataDTO(val inventorySpace: Int, val day: Int, val money: Int, val inventoryItems: MutableList<Pair<Item, Int>>)
+
