+++ src/main/kotlin/com/projects/inGameMarketplace/itemService/ItemRepository.kt
@@ -20,7 +20,7 @@ class ItemRepository {
         while (resultSet.next()) {
             val name = resultSet.getString("item_name")
             val image = resultSet.getString("image_url")
-            val price = resultSet.getDouble("price")
+            val price = resultSet.getInt("price")
 
             val entity = ItemEntity(name, image, price)
             listBuilder.add(entity)
